major refactoring of base agent code structure to make it easier to work with
